Extract review packet delivery contract
